MEMORANDUM
Serenity on the Lakes: A Self-sustained Community that offers farm to table, cruelty-free, eco- friendly Urbanization that embraces the necessity for advanced technology and Rehabilitation for its residence. With an Emphasis on Veterans, Ex-Offenders and the Growing Elderly Populations we will evaluate the missing Opportunity link in our society that capitalizes on their vulnerabilities for profit. Profit through overpricing and insufficient health care under- represented neighborhoods, Recidivism due to lack of opportunities and support and delayed bureaucracy for Military related mental illnesses.
From Concept to Realization, This Memorandum will explore the mechanics of breathing life into a community which nurtures, teaches and supports its residents in preparation for a technologically driven world by removing the intimidation factor. Through an actually planned community under consideration by the City of Houston. Funding Contractual Negotiations, A visualization will be explained and critiqued on the necessities of building such communities in preparation for the future.